+++ CMakeLists.txt
@@ -479,6 +479,7 @@ set(COMPLEX_SRCS
   ${COMPLEX_SOURCE_DIR}/Utilities/ParallelTaskAlgorithm.cpp
   ${COMPLEX_SOURCE_DIR}/Utilities/SegmentFeatures.cpp
   ${COMPLEX_SOURCE_DIR}/Utilities/AlignSections.cpp
+  ${COMPLEX_SOURCE_DIR}/Utilities/OStreamUtilities.cpp
 
   ${COMPLEX_SOURCE_DIR}/Utilities/Math/GeometryMath.cpp
   ${COMPLEX_SOURCE_DIR}/Utilities/Math/MatrixMath.cpp
